// src/app/models/users.ts
export class User {
    user_Unique_Id_Number: number;
    user_name: string;
    user_email: string;
    user_login: string;
    user_password: string;
    user_password_modification_required: boolean;
    user_date_of_creation: Date;
    user_is_active: boolean;
    user_phone_professional: string;
    user_phone_Mobile1: string;
    user_phone_Mobile2: string;
    user_fax: string;
    user_picture: string;
    user_creator: string;

}